refactor(core): break authority summary import cycle
